--- packages/insomnia/src/ui/components/websockets/types.ts
@@ -1,6 +1,0 @@
-export enum ReadyState {
-  CONNECTING = 0,
-  OPEN = 1,
-  CLOSING = 2,
-  CLOSED = 3,
-}
