+++ packages/insomnia/src/ui/context/websocket-client/create-websocket-client.ts
@@ -1,0 +1,14 @@
+export interface WebSocketClient {
+    create: typeof window.main.webSocketConnection.create;
+    close: typeof window.main.webSocketConnection.close;
+    send: typeof window.main.webSocketConnection.event.send;
+    onReadyState: typeof window.main.webSocketConnection.readyState.subscribe;
+  }
+export function createWebSocketClient(): WebSocketClient {
+  return {
+    create: window.main.webSocketConnection.create,
+    close: window.main.webSocketConnection.close,
+    send: window.main.webSocketConnection.event.send,
+    onReadyState: window.main.webSocketConnection.readyState.subscribe,
+  };
+}
