+++ packages/insomnia/src/ui/context/websocket-client/use-ws-message.ts
@@ -2,19 +2,20 @@ import { useEffect, useState } from 'react';
 
 import { WebsocketEvent } from '../../../main/network/websocket';
 
-export function useWSEvent(requestId: string): WebsocketEvent | null {
-  const [event, setEvent] = useState<WebsocketEvent | null>(null);
+// this hook is only for development purpose
+export function useWSMessage(requestId: string): WebsocketEvent | null {
+  const [message, setMessage] = useState<WebsocketEvent | null>(null);
 
   useEffect(() => {
     const unsubscribe = window.main.webSocketConnection.event.subscribe(
       { requestId },
       (incomingEvent: WebsocketEvent) => {
-        setEvent(incomingEvent);
+        setMessage(incomingEvent);
       }
     );
 
     return unsubscribe;
   }, [requestId]);
 
-  return event;
+  return message;
 }
