+++ packages/insomnia/src/ui/context/websocket-client/websocket-client-context.tsx
@@ -1,19 +1,7 @@
 import React, { createContext, FC, ReactNode, useContext } from 'react';
 
-interface WebSocketClient {
-  create: typeof window.main.webSocketConnection.create;
-  close: typeof window.main.webSocketConnection.close;
-  send: typeof window.main.webSocketConnection.event.send;
-  onReadyState: typeof window.main.webSocketConnection.readyState.subscribe;
-}
-export function createWebSocketClient(): WebSocketClient {
-  return {
-    create: window.main.webSocketConnection.create,
-    close: window.main.webSocketConnection.close,
-    send: window.main.webSocketConnection.event.send,
-    onReadyState: window.main.webSocketConnection.readyState.subscribe,
-  };
-}
+import { WebSocketClient } from './create-websocket-client';
+
 export const WebSocketClientContext = createContext<WebSocketClient | undefined>(undefined);
 interface Props {
   client: WebSocketClient;
